+++ CHANGELOG.md
@@ -2291,3 +2291,4 @@
 - [2026-06-25 09:32] fix(cache): prevent double tap duplicate action in cache
 - [2026-06-25 09:54] test(state): add unit tests for state business logic
 - [2026-06-25 10:28] fix(state): prevent double tap duplicate action in state
+- [2026-06-25 10:46] refactor(ui): decouple repository interface from concrete ui service